+++ go-flutter/chapter-3/go.mod
@@ -1,9 +1,12 @@
 module howa.in/chapter-3
 
-go 1.24.1
+go 1.24.2
 
-toolchain go1.24.2
+require github.com/youmark/pkcs8 v0.0.0-20240726163527-a2c0da244d78 // indirect
 
-require github.com/youmark/pkcs8 v0.0.0-20240726163527-a2c0da244d78
+require (
+	golang.org/x/crypto v0.38.0 // indirect
+	howa.in/common v0.0.0-00010101000000-000000000000
+)
 
-require golang.org/x/crypto v0.36.0 // indirect
+replace howa.in/common => ../common
