+++ go-flutter/chapter-4/finance/finance.go
@@ -26,83 +26,19 @@ package main
 
 import (
 	"context"
-	"crypto/rsa"
 	"crypto/tls"
-	"crypto/x509"
 	"encoding/json"
-	"encoding/pem"
 	"fmt"
 	"log"
 	"net/http"
 	"net/url"
-	"os"
 
 	"github.com/crewjam/saml/samlsp"
-	"github.com/youmark/pkcs8"
+	"howa.in/common"
 )
 
-func getProviderCertAndKey(certpath, keypath string, keypass []byte) (key *rsa.PrivateKey, cert *x509.Certificate, err error) {
-	var data []byte
-	if data, err = os.ReadFile(keypath); err == nil {
-		if block, _ := pem.Decode(data); block != nil {
-			if key, err = pkcs8.ParsePKCS8PrivateKeyRSA(block.Bytes, keypass); err != nil {
-				return
-			}
-		}
-	}
-	if data, err = os.ReadFile(certpath); err == nil {
-		if block, _ := pem.Decode(data); block != nil {
-			cert, err = x509.ParseCertificate(block.Bytes)
-		}
-	}
-	return
-}
-
-func addCertificates(certpath string, c *tls.Certificate) (err error) {
-	var (
-		data  []byte
-		block *pem.Block
-	)
-	if data, err = os.ReadFile(certpath); err == nil {
-		for block, data = pem.Decode(data); block != nil; block, data = pem.Decode(data) {
-			if block.Type == "CERTIFICATE" {
-				c.Certificate = append(c.Certificate, block.Bytes)
-			}
-		}
-	}
-	return
-}
-
-/*
-Server certificate
-*/
-func getTLSCert(capath, certpath, keypath string, keypass []byte) (c *tls.Certificate, err error) {
-	var (
-		data  []byte
-		block *pem.Block
-		cert  tls.Certificate
-	)
-
-	if err = addCertificates(certpath, &cert); err == nil {
-		if err = addCertificates(capath, &cert); err == nil {
-			if data, err = os.ReadFile(keypath); err == nil {
-				if block, _ = pem.Decode(data); block != nil {
-					if cert.PrivateKey, _, err = pkcs8.ParsePrivateKey(block.Bytes, keypass); err == nil {
-						if cert.Leaf, err = x509.ParseCertificate(cert.Certificate[0]); err == nil {
-							c = &cert
-						}
-					}
-				} else {
-					err = fmt.Errorf("no private key data found")
-				}
-			}
-		}
-	}
-	return
-}
-
 func setupTLSServer(srvName string) *http.Server {
-	cert, err := getTLSCert(
+	cert, err := common.GetTLSCert(
 		"certs/ssl/scas.crt",
 		fmt.Sprintf("certs/ssl/%s.crt", srvName),
 		fmt.Sprintf("certs/ssl/%s.key", srvName),
@@ -123,31 +59,6 @@ func setupTLSServer(srvName string) *http.Server {
 	}
 }
 
-func getHTTPSClient(capath string) (client *http.Client, err error) {
-	var (
-		tlsConfig tls.Config
-		data      []byte
-	)
-	if data, err = os.ReadFile(capath); err == nil {
-		var block *pem.Block
-		certpool := x509.NewCertPool()
-		for block, data = pem.Decode(data); block != nil; block, data = pem.Decode(data) {
-			if cert, err := x509.ParseCertificate(block.Bytes); err == nil {
-				certpool.AddCert(cert)
-			}
-		}
-		tlsConfig.RootCAs = certpool
-	} else {
-		return
-	}
-	client = &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tlsConfig,
-		},
-	}
-	return
-}
-
 func main() {
 	rootURL, err := url.Parse("https://finance.mysrv.local:8445")
 	if err != nil {
@@ -159,7 +70,7 @@ func main() {
 		log.Default().Fatalf("IDP not running at: https://idp.local:8443 %v", err)
 	}
 
-	httpclient, err := getHTTPSClient("certs/ssl/scas.crt")
+	httpclient, err := common.GetHTTPSClient("certs/ssl/scas.crt")
 	if err != nil {
 		log.Default().Fatalf("Unable to read RootCAs: %v", err)
 	}
@@ -169,7 +80,7 @@ func main() {
 		log.Default().Fatalf("Failed to download IDP metadata %v", err)
 	}
 
-	key, cert, err := getProviderCertAndKey("certs/finance.crt", "certs/finance.key", []byte("password"))
+	key, cert, err := common.GetProviderCertAndKey("certs/finance.crt", "certs/finance.key", []byte("password"))
 	if err != nil {
 		panic(err)
 	}
