--- go-flutter/chapter-5/certs/idp.local.p12
@@ -1,1 +1,0 @@
-../../appendix-c/output/mysrv.local.p12
