--- go-flutter/chapter-5/certs/mysrv.p12
@@ -1,1 +1,0 @@
-../../appendix-c/output/mysrv.local.p12
