+++ go-flutter/chapter-5/pkce/client/pubspec.yaml
@@ -36,7 +36,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   url_launcher: ^6.1.11
   provider: ^6.0.5
-  http: ^0.13.6
+  http: ^1.4.0
   crypto: ^3.0.3
   oauth2: ^2.0.1
 
@@ -49,7 +49,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^5.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
