+++ go-flutter/chapter-5/pkce/resource/go.mod
@@ -1,5 +1,11 @@
 module server
 
-go 1.20
+go 1.24.2
 
-require golang.org/x/crypto v0.9.0
+require github.com/youmark/pkcs8 v0.0.0-20240726163527-a2c0da244d78 // indirect
+
+require golang.org/x/crypto v0.38.0 // indirect
+
+require howa.in/common v0.0.0-00010101000000-000000000000
+
+replace howa.in/common => ../../../common
