+++ go-flutter/chapter-5/pkce/resource/server.go
@@ -25,7 +25,6 @@ package main
 
 import (
 	"crypto/tls"
-	"encoding/pem"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -34,28 +33,15 @@ import (
 	"net/http/httputil"
 	"os"
 
-	"github.com/go-openssl/pkcs12"
+	"howa.in/common"
 )
 
-func getTLSCert(certloc string) (cert tls.Certificate, err error) {
-	var (
-		fdata   []byte
-		blocks  []*pem.Block
-		pemData []byte
-	)
-	if fdata, err = os.ReadFile(certloc); err == nil {
-		if blocks, err = pkcs12.ToPEM(fdata, "password"); err == nil {
-			for _, b := range blocks {
-				pemData = append(pemData, pem.EncodeToMemory(b)...)
-			}
-			cert, err = tls.X509KeyPair(pemData, pemData)
-		}
-	}
-	return
-}
-
-func setupTLSServer(certloc string, srvName string) *http.Server {
-	cert, err := getTLSCert(certloc)
+func setupTLSServer(srvName string) *http.Server {
+	cert, err := common.GetTLSCert(
+		"../../certs/scas.crt",
+		fmt.Sprintf("../../certs/%s.crt", srvName),
+		fmt.Sprintf("../../certs/%s.key", srvName),
+		[]byte("password"))
 	if err != nil {
 		log.Default().Fatal(err)
 	}
@@ -63,7 +49,7 @@ func setupTLSServer(certloc string, srvName string) *http.Server {
 	tlsConfig := &tls.Config{
 		ServerName:   srvName,
 		MinVersion:   tls.VersionTLS13,
-		Certificates: []tls.Certificate{cert},
+		Certificates: []tls.Certificate{*cert},
 	}
 
 	return &http.Server{
@@ -121,7 +107,7 @@ func main() {
 		}
 	})
 
-	server := setupTLSServer("../../certs/mysrv.p12", "mysrv.local")
+	server := setupTLSServer("mysrv.local")
 	log.Fatal(server.ListenAndServeTLS("", ""))
 }
