+++ go-flutter/common/go.mod
@@ -1,0 +1,7 @@
+module howa.in/common
+
+go 1.24.2
+
+require github.com/youmark/pkcs8 v0.0.0-20240726163527-a2c0da244d78
+
+require golang.org/x/crypto v0.22.0 // indirect
